api/pyproject.toml
[project]
name = "ragify"
version = "0.1.0"
description = "Adaptive RAG"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.135.2",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.12",
    "langchain-qdrant>=1.1.0",
    "langgraph>=1.1.3",
    "markitdown>=0.1.5",
    "pydotenv>=0.0.7",
    "pyjwt>=2.12.1",
    "pymongo[srv]>=4.16.0",
    "qdrant-client>=1.17.1",
]
